requirements.txt
PyPDF2==3.0.1
google-generativeai>=0.8.0
pathlib
Flask==3.0.0
Flask-CORS==4.0.0
requests>=2.31.0
